+++ CHANGELOG.md
@@ -1535,7 +1535,7 @@ module.exports = {
             level: "info",
             // Using colors on the output
             colors: true,
-            // Print module names with different colors (like docker-compose for containers)
+            // Print module names with different colors (like docker compose for containers)
             moduleColors: false,
             // Line formatter. It can be "json", "short", "simple", "full", a `Function` or a template string like "{timestamp} {level} {nodeID}/{mod}: {msg}"
             formatter: "full",